+++ src/panels/SectionElementPanel.js
@@ -42,13 +42,6 @@ export default class SectionElementPanel {
                         elDataSource.val(), SetValueCmd.type.text, this.rb);
                     this.rb.executeCommand(cmd);
                 }
-            })
-            .blur(event => {
-                this.rb.getPopupWindow().hide();
-            })
-            .mouseup(event => {
-                event.preventDefault();
-                event.stopPropagation();
             });
         autosize(elDataSource);
         elFormField.append(elDataSource);